Add feedback after event created
Creating an event didnt have any feedback on earlier versions if your event had been created or if the button did anything, so i made a pop-up after one has created an event. it asks if the user wants to either go to the hompage (where the new event now is located) or to make another one.